+++ stapi-pydantic/CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - pydantic >= 2.12 is now required.
 
+### Changed
+
+- pydantic >= 2.12 is now required.
+
 ## [0.0.4] - 2025-07-17
 
 ### Added